+++ jackson/src/main/java/veslo/JacksonTestClient.java
@@ -1,0 +1,60 @@
+/*
+ * Copyright 2021-2022 Shaburov Oleg
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package veslo;
+
+import okhttp3.Interceptor;
+import retrofit2.CallAdapter;
+import retrofit2.internal.EverythingIsNonNull;
+import veslo.client.TestClient;
+import veslo.client.adapter.UniversalCallAdapterFactory;
+import veslo.client.inteceptor.CompositeInterceptor;
+
+/**
+ * Utility class for building an HTTP client with a built-in Jackson2 converter
+ * <p>
+ *
+ * @author Oleg Shaburov (shaburov.o.a@gmail.com)
+ * Created: 04.02.2022
+ */
+public class JacksonTestClient {
+
+    /**
+     * Builds an HTTP client with a built-in Jackson2 converter
+     *
+     * @param baseUrl            - HTTP resource URL
+     * @param interceptor        - {@link CompositeInterceptor} (okhttp network interceptor)
+     * @param callAdapterFactory - {@link UniversalCallAdapterFactory} or heirs
+     * @param cli                - client interface class
+     * @param <CLIENT>           - client interface
+     * @return built client
+     */
+    @EverythingIsNonNull
+    public static <CLIENT> CLIENT build(final String baseUrl,
+                                        final Interceptor interceptor,
+                                        final CallAdapter.Factory callAdapterFactory,
+                                        final Class<CLIENT> cli) {
+        return TestClient.build(baseUrl, interceptor, callAdapterFactory, new JacksonConverterFactory(), cli);
+    }
+
+    /**
+     * Utility class. Forbidden instantiation.
+     */
+    private JacksonTestClient() {
+        throw new UtilityClassException();
+    }
+
+}