+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Change Log
 ==========
 
+## Version 1.0.1
+
+* **New** utility classes for building test clients: 
+  `TestClient`, `JacksonTestClient`, `GsonTestClient`, `Veslo4Test`
+* **Change** packaging for `all` module changed from `pom` to `jar`
+* **Update** up slf4j-api version from 1.7.32 to 1.7.35
+
 ## Version 1.0.0
 
 * **New**: `JavaTypeCallAdapterFactory` to convert the response body to primitive or reference java types